least priviledge + some small fixes

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -31,16 +31,12 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try
+        if ($this->isAble('store', Ticket::class))
         {
-            $this->isAble('store', Ticket::class);
-
             return new TicketResource(Ticket::create($request->mappedAttributes()));
         }
-        catch (AuthorizationException $e)
-        {
-            return $this->error('You are not authorised to create that resource', 401);
-        }
+
+        return $this->error('You are not authorised to create that resource', 401);
     }
 
     /**
@@ -75,20 +71,18 @@ class TicketController extends ApiController
         {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            // Policy
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $table))
+            {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not authorised to update that resource', 401);
         }
         catch (ModelNotFoundException $e)
         {
             return $this->error('Ticket cannot be found', 404);
-        }
-        catch (AuthorizationException $e)
-        {
-            return $this->error('You are not authorised to update that resource', 401);
         }
     }
 
@@ -99,19 +93,18 @@ class TicketController extends ApiController
         {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket))
+            {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not authorised to update that resource', 401);
         }
         catch (ModelNotFoundException $e)
         {
             return $this->error('Ticket cannot be found', 404);
-        }
-        catch (AuthorizationException $e)
-        {
-            return $this->error('You are not authorised to update that resource', 401);
         }
     }
 
@@ -124,19 +117,18 @@ class TicketController extends ApiController
         {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('delete', $ticket);
+            if ($this->isAble('delete', $ticket))
+            {
+                $ticket->delete();
 
-            $ticket->delete();
+                return $this->ok('Ticket deleted successfully');
+            }
 
-            return $this->ok('Ticket deleted successfully');
+            return $this->error('You are not authorised to delete that resource', 401);
         }
         catch (ModelNotFoundException $e)
         {
             return $this->error('Ticket cannot be found', 404);
-        }
-        catch (AuthorizationException $e)
-        {
-            return $this->error('You are not authorised to delete that resource', 401);
         }
     }
 }
